requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy-serializer==1.4.1
Faker==22.0.0
python-dateutil==2.8.2 